Moving everything to examples since its not specific to sotopia
